requirements.txt
requests
numpy
scikit-learn
sentence-transformers
openai
python-dotenv
streamlit

